python-workers/requirements.txt
torch==2.0.1
torchvision==0.15.2
torchaudio==2.0.2
transformers==4.35.0
openai-whisper==20231117
pyannote.audio==3.1.0
librosa==0.10.0
numpy==1.24.3
pandas==2.0.3
soundfile==0.12.1
ffmpeg-python==0.2.0
python-dotenv==1.0.0
jupyter==1.0.0